+++ packages/walletkit/src/hooks/useWalletKitSwitchNetwork.ts
@@ -3,8 +3,11 @@ import { useSwitchNetwork } from 'wagmi';
 import { useWalletKitContext } from '..';
 
 export type UseWalletKitSwitchNetworkProps = Parameters<typeof useSwitchNetwork>[0];
+export type UseWalletKitSwitchNetworkReturnType = ReturnType<typeof useSwitchNetwork>;
 
-export function useWalletKitSwitchNetwork(props?: UseWalletKitSwitchNetworkProps) {
+export function useWalletKitSwitchNetwork(
+  props?: UseWalletKitSwitchNetworkProps,
+): UseWalletKitSwitchNetworkReturnType {
   const { log, options } = useWalletKitContext();
 
   const result = useSwitchNetwork({