+++ packages/walletkit/src/wallets/index.ts
@@ -14,3 +14,6 @@ export * from './okxWallet';
 export * from './mathWallet';
 export * from './binanceWeb3Wallet';
 export * from './coinbaseWallet';
+export * from './coinbaseWallet/connector';
+// export * from './particleWallet';
+// export * from './particleWallet/connector';