+++ packages/walletkit/test/vite.config.ts
@@ -5,7 +5,8 @@ import react from '@vitejs/plugin-react';
 import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
 
 export default defineConfig({
-  root: './dev',
+  envDir: './',
+  root: './test',
   plugins: [
     react(),
     vanillaExtractPlugin({