+++ packages/walletkit/CHANGELOG.md
@@ -1,4 +1,34 @@
-# @totejs/walletkit
+# @node-real/walletkit
+
+## 1.0.9-alpha.4
+
+### Patch Changes
+
+- e06d4b3: refactor: Add env variables.
+
+## 1.0.9-alpha.3
+
+### Patch Changes
+
+- a9a34af: Add particle wallet
+
+## 1.0.9-alpha.2
+
+### Patch Changes
+
+- a6de228: Add `closeModalAfterSwitchingNetwork`.
+
+## 1.0.9-alpha.1
+
+### Patch Changes
+
+- f3ebc1b: Setting `reloadOnDisconnect` = false in default for CoinbaseWallet.
+
+## 1.0.9-alpha.0
+
+### Patch Changes
+
+- 53b65ec: Fixed switching network failure in coinbaseWallet.
 
 ## 1.0.8